+++ packages/typescript-config/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @repo/typescript-config
 
+## 0.1.0
+
+### Minor Changes
+
+- 3ba4563: beta version
+
+### Patch Changes
+
+- c5cb745: fix cli
+- 2b9ab3c: update cli package
+- ea94003: quick fix
+- 4c18a70: update recipes
+- 29af253: Update CLI package
+- 65c9202: change component import resolution
+
 ## 0.1.0-beta.6
 
 ### Patch Changes